--- a/application/models/models_test.go
+++ b/application/models/models_test.go
@@ -5,24 +5,69 @@ import (
 	"testing"
 
 	"github.com/gobuffalo/buffalo"
+	"github.com/gobuffalo/envy"
 	"github.com/gobuffalo/nulls"
-	"github.com/gobuffalo/suite/v3"
+	"github.com/gobuffalo/pop/v5"
 	"github.com/gofrs/uuid"
+	"github.com/stretchr/testify/require"
+	"github.com/stretchr/testify/suite"
 
 	"github.com/silinternational/wecarry-api/domain"
 )
 
+// ModelSuite doesn't contain a buffalo suite.Model and can be used for tests that don't need access to the database
+// or don't need the buffalo test runner to refresh the database
 type ModelSuite struct {
-	*suite.Model
+	suite.Suite
+	*require.Assertions
+	DB *pop.Connection
 }
 
-func Test_ModelSuite(t *testing.T) {
-	model := suite.NewModel()
+func (m *ModelSuite) SetupTest() {
+	m.Assertions = require.New(m.T())
+	m.DestroyAll()
+}
 
-	as := &ModelSuite{
-		Model: model,
+func (m *ModelSuite) DestroyAll() {
+	// delete all Requests, RequestHistories, RequestFiles, PotentialProviders, Threads, and ThreadParticipants
+	var requests Requests
+	m.NoError(m.DB.All(&requests))
+	m.NoError(m.DB.Destroy(&requests))
+
+	// delete all Meetings, MeetingParticipants, and MeetingInvites
+	var meetings Meetings
+	m.NoError(m.DB.All(&meetings))
+	m.NoError(m.DB.Destroy(&meetings))
+
+	// delete all Organizations, OrganizationDomains, OrganizationTrusts, and UserOrganizations
+	var organizations Organizations
+	m.NoError(m.DB.All(&organizations))
+	m.NoError(m.DB.Destroy(&organizations))
+
+	// delete all Users, Messages, UserAccessTokens, and Watches
+	var users Users
+	m.NoError(m.DB.All(&users))
+	m.NoError(m.DB.Destroy(&users))
+
+	// delete all Files
+	var files Files
+	m.NoError(m.DB.All(&files))
+	m.NoError(m.DB.Destroy(&files))
+
+	// delete all Locations
+	var locations Locations
+	m.NoError(m.DB.All(&locations))
+	m.NoError(m.DB.Destroy(&locations))
+}
+
+// Test_ModelSuite runs the test suite
+func Test_ModelSuite(t *testing.T) {
+	ms := &ModelSuite{}
+	c, err := pop.Connect(envy.Get("GO_ENV", "test"))
+	if err == nil {
+		ms.DB = c
 	}
-	suite.Run(t, as)
+	suite.Run(t, ms)
 }
 
 func createFixture(ms *ModelSuite, f interface{}) {