Use custom test suites in `actions` and `models`. On my machine, it reduced test time from 103 seconds to 14 seconds on `actions` and from 386 seconds to 31 seconds on `models`. In the process, found a test issue that may come back to bite later. Should probably redesign the test fixture creation to look a bit more like TF. Specifically, pass users into request fixture generation functions to make sure that requests (etc.) are actually created by that user.